refactor(taskflow): move data/db/util under infra
Relocates all hand-written infra code under the new `infra` package:
- data/local, data/remote, data/sync → infra/data/{local,remote,sync}
- data/SeedData → infra/SeedData
- db/Adapters.kt → infra/db/Adapters.kt (adds generated db.* imports)
- util/IdGenerator.kt → infra/util/IdGenerator.kt
- Corresponding test files moved to matching infra/… paths

Updates package declarations and all import sites tree-wide.
Generated SQLDelight files under db/ stay in place.
