+++ examples/taskflow/composeApp/src/commonMain/kotlin/org/reduxkotlin/sample/taskflow/App.kt
@@ -45,10 +45,11 @@ import org.reduxkotlin.sample.taskflow.core.AppSettingsModel
 import org.reduxkotlin.sample.taskflow.core.BoardId
 import org.reduxkotlin.sample.taskflow.core.NavModel
 import org.reduxkotlin.sample.taskflow.core.Route
-import org.reduxkotlin.sample.taskflow.data.SeedData
-import org.reduxkotlin.sample.taskflow.data.local.LocalStore
-import org.reduxkotlin.sample.taskflow.data.local.SqlDelightLocalStore
-import org.reduxkotlin.sample.taskflow.db.taskFlowDb
+import org.reduxkotlin.sample.taskflow.infra.SeedData
+import org.reduxkotlin.sample.taskflow.infra.data.local.LocalStore
+import org.reduxkotlin.sample.taskflow.infra.data.local.SqlDelightLocalStore
+import org.reduxkotlin.sample.taskflow.infra.db.taskFlowDb
+import org.reduxkotlin.sample.taskflow.infra.util.DefaultIdGenerator
 import org.reduxkotlin.sample.taskflow.model.AccountsModel
 import org.reduxkotlin.sample.taskflow.model.AuthMode
 import org.reduxkotlin.sample.taskflow.platform.DriverFactory
@@ -71,7 +72,6 @@ import org.reduxkotlin.sample.taskflow.ui.screens.ProfileScreen
 import org.reduxkotlin.sample.taskflow.ui.screens.SettingsScreen
 import org.reduxkotlin.sample.taskflow.ui.screens.SwitcherScreen
 import org.reduxkotlin.sample.taskflow.ui.theme.TaskFlowTheme
-import org.reduxkotlin.sample.taskflow.util.DefaultIdGenerator
 import kotlin.time.Clock
 
 /**
